Migrate BAM I/O from rust-htslib to noodles (#115)
* Migrate BAM I/O from rust-htslib to noodles (#113)

Replace rust-htslib with a pure-Rust noodles backend behind a
compatibility shim at src/rna/bam/. The shim preserves the existing
Reader/IndexedReader/Record API used across dupRadar, RSeQC, preseq,
Qualimap, and samtools-compatible outputs.

Key changes:
- Add noodles-backed BAM/SAM/CRAM readers with indexed fetch support
- Preserve samtools-identical CHK checksums via packed sequence bytes
- Update build docs: no cmake/htslib system deps required
- All integration and unit tests pass in release mode


* Update docs, Dockerfile, and CI for noodles migration

Replace rust-htslib build prerequisites with noodles (pure Rust) across
the website docs, Dockerfile, GitHub Actions workflows, and CHANGELOG.


* Fix CI: rustfmt and bump MSRV to 1.89 for noodles

- Remove extra blank line in main.rs (cargo fmt --check)
- Bump rust-version to 1.89 (required by noodles 0.111)
- Update MSRV CI job to match


---------

+++ src/rna/preseq.rs
@@ -4,9 +4,9 @@
 //! sequencing depth using the Good-Toulmin rational function extrapolation
 //! method, matching the behavior of preseq v3.
 
+use crate::rna::bam;
 use anyhow::{bail, Context, Result};
 use log::debug;
-use rust_htslib::bam;
 use std::collections::HashMap;
 use std::io::Write;
 use std::path::Path;
@@ -154,7 +154,7 @@ impl PreseqAccum {
 
         let tid = record.tid();
         let start = record.pos();
-        let end = record.cigar().end_pos();
+        let end = record.cigar().end_pos() as i64;
 
         let info = MateInfo { tid, start, end };
